+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/HadoopFileSystemTestBase.java
@@ -25,6 +25,7 @@ import static com.google.cloud.hadoop.fs.gcs.GhfsStatistic.STREAM_READ_SEEK_BYTE
 import static com.google.cloud.hadoop.fs.gcs.GhfsStatistic.STREAM_READ_SEEK_FORWARD_OPERATIONS;
 import static com.google.cloud.hadoop.fs.gcs.GhfsStatistic.STREAM_READ_SEEK_OPERATIONS;
 import static com.google.cloud.hadoop.fs.gcs.GhfsStatistic.STREAM_READ_TOTAL_BYTES;
+import static com.google.cloud.hadoop.fs.gcs.GhfsStatistic.STREAM_READ_VECTORED_OPERATIONS;
 import static com.google.cloud.hadoop.fs.gcs.GhfsStatistic.STREAM_WRITE_BYTES;
 import static com.google.common.base.Strings.isNullOrEmpty;
 import static com.google.common.truth.Truth.assertThat;
@@ -37,9 +38,11 @@ import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystemIntegrationTest
 import com.google.cloud.hadoop.gcsio.StringPaths;
 import com.google.common.collect.ImmutableList;
 import com.google.common.flogger.GoogleLogger;
+import java.io.EOFException;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.URI;
+import java.nio.ByteBuffer;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -47,13 +50,20 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.*;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.TimeUnit;
 import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FSDataOutputStream;
+import org.apache.hadoop.fs.FileRange;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.StorageStatistics;
+import org.apache.hadoop.util.functional.FutureIO;
 import org.junit.Test;
 
 /**
@@ -71,6 +81,8 @@ public abstract class HadoopFileSystemTestBase extends GoogleCloudStorageFileSys
   FileSystem ghfs;
 
   protected HadoopFileSystemIntegrationHelper ghfsHelper;
+  /** Timeout in seconds for vectored read operation in tests : {@value}. */
+  public static final int VECTORED_READ_OPERATION_TEST_TIMEOUT_SECONDS = 5 * 60;
 
   @Override
   public void postCreateInit() throws IOException {
@@ -906,6 +918,33 @@ public abstract class HadoopFileSystemTestBase extends GoogleCloudStorageFileSys
   }
 
   @Test
+  public void testInputStreamReadVectoredIOStatistics() throws Exception {
+
+    // Write an object.
+    URI path = getTempFilePath();
+    Path hadoopPath = ghfsHelper.castAsHadoopPath(path);
+    String text = "Hello World!";
+    int numBytesWritten = ghfsHelper.writeFile(hadoopPath, text, 1024, /* overwrite= */ false);
+
+    try (FSDataInputStream readStream = ghfs.open(hadoopPath)) {
+
+      List<FileRange> fileRanges = new ArrayList<>();
+      fileRanges.add(FileRange.createFileRange(0, 10));
+      readStream.readVectored(fileRanges, ByteBuffer::allocate);
+      validateVectoredReadResult(fileRanges, hadoopPath);
+
+      assertThat(
+              readStream
+                  .getIOStatistics()
+                  .counters()
+                  .get(STREAM_READ_VECTORED_OPERATIONS.getSymbol()))
+          .isEqualTo(1);
+    } finally {
+      ghfs.delete(hadoopPath);
+    }
+  }
+
+  @Test
   public void testInputStreamSeekIOStatistics() throws IOException {
 
     // Write an object.
@@ -989,6 +1028,58 @@ public abstract class HadoopFileSystemTestBase extends GoogleCloudStorageFileSys
     assertThat(ghfs.delete(hadoopPath)).isTrue();
   }
 
+  @Test
+  public void testSomeRangesMergedSomeUnmerged() throws Exception {
+    Path hadoopPath = ghfsHelper.castAsHadoopPath(getTempFilePath());
+    byte[] testBytes = new byte[8 * 8 * 1024];
+    for (int i = 0; i < testBytes.length; i++) {
+      testBytes[i] = (byte) (i * i);
+    }
+
+    ghfsHelper.writeFile(hadoopPath, testBytes, 1, /* overwrite= */ false);
+
+    List<FileRange> fileRanges = new ArrayList<>();
+    fileRanges.add(FileRange.createFileRange(8 * 1024, 100));
+    fileRanges.add(FileRange.createFileRange(14 * 1024, 100));
+    fileRanges.add(FileRange.createFileRange(10 * 1024, 100));
+    fileRanges.add(FileRange.createFileRange(2 * 1024 - 101, 100));
+    fileRanges.add(FileRange.createFileRange(40 * 1024, 1024));
+    try (FSDataInputStream input = ghfs.open(hadoopPath)) {
+      input.readVectored(fileRanges, ByteBuffer::allocate);
+      validateVectoredReadResult(fileRanges, hadoopPath);
+    } finally {
+      ghfs.delete(hadoopPath);
+    }
+  }
+
+  @Test
+  public void testInvalidRangeRequest() throws Exception {
+    Path hadoopPath = ghfsHelper.castAsHadoopPath(getTempFilePath());
+    byte[] testBytes = new byte[8 * 8 * 1024];
+    for (int i = 0; i < testBytes.length; i++) {
+      testBytes[i] = (byte) (i * i);
+    }
+
+    ghfsHelper.writeFile(hadoopPath, testBytes, 1, /* overwrite= */ false);
+    List<FileRange> fileRanges = new ArrayList<>();
+    fileRanges.add(FileRange.createFileRange(-1, 10));
+    try (FSDataInputStream input = ghfs.open(hadoopPath)) {
+      assertThrows(EOFException.class, () -> input.readVectored(fileRanges, ByteBuffer::allocate));
+    } finally {
+      ghfs.delete(hadoopPath);
+    }
+  }
+
+  private void validateVectoredReadResult(List<FileRange> fileRanges, Path hadoopPath)
+      throws Exception {
+    for (FileRange range : fileRanges) {
+      ByteBuffer resBuffer =
+          FutureIO.awaitFuture(
+              range.getData(), VECTORED_READ_OPERATION_TEST_TIMEOUT_SECONDS, TimeUnit.SECONDS);
+      GoogleHadoopFileSystemTestHelper.assertObjectContent(
+          ghfs, hadoopPath, resBuffer.duplicate(), range.getOffset());
+    }
+  }
   // -----------------------------------------------------------------
   // Inherited tests that we suppress because they do not make sense
   // in the context of this layer.